+++ src/types/integer.rs
@@ -1,5 +1,5 @@
+use crate::element::owned::Element;
 use crate::result::{decoding_error, IonError};
-use crate::value::owned::Element;
 use num_bigint::{BigInt, BigUint, ToBigUint};
 use num_traits::{ToPrimitive, Zero};
 use std::cmp::Ordering;
@@ -12,8 +12,8 @@ pub trait IntAccess {
     ///
     /// ## Usage
     /// ```
-    /// # use ion_rs::value::*;
-    /// # use ion_rs::value::owned::*;
+    /// # use ion_rs::element::*;
+    /// # use ion_rs::element::owned::*;
     /// # use ion_rs::types::integer::*;
     /// # use num_bigint::*;
     /// let big_int = Integer::BigInt(BigInt::from(100));
@@ -35,8 +35,8 @@ pub trait IntAccess {
     ///
     /// ## Usage
     /// ```
-    /// # use ion_rs::value::*;
-    /// # use ion_rs::value::owned::*;
+    /// # use ion_rs::element::*;
+    /// # use ion_rs::element::owned::*;
     /// # use ion_rs::types::integer::*;
     /// # use num_bigint::*;
     /// # use std::str::FromStr;
